Add additional tests for top level await
