+++ test/e2e/async-modules-app/app/app-router/client-component/page.tsx
@@ -1,0 +1,6 @@
+'use client'
+const appValue = await Promise.resolve('hello')
+
+export default function Page() {
+  return <p id="app-router-client-component-value">{appValue}</p>
+}
