+++ test/e2e/async-modules-app/app/app-router/page.tsx
@@ -1,0 +1,5 @@
+const appValue = await Promise.resolve('hello')
+
+export default function Page() {
+  return <p id="app-router-value">{appValue}</p>
+}
