+++ test/e2e/async-modules-app/index.test.ts
@@ -1,0 +1,19 @@
+/* eslint-env jest */
+import { nextTestSetup } from 'e2e-utils'
+
+describe('Async modules', () => {
+  const { next } = nextTestSetup({
+    files: __dirname,
+  })
+  it('app router server component async module', async () => {
+    const browser = await next.browser('/app-router')
+    expect(await browser.elementByCss('#app-router-value').text()).toBe('hello')
+  })
+
+  it('app router client component async module', async () => {
+    const browser = await next.browser('/app-router/client')
+    expect(
+      await browser.elementByCss('#app-router-client-component-value').text()
+    ).toBe('hello')
+  })
+})
