+++ test/e2e/async-modules/next.config.js
@@ -1,0 +1,7 @@
+module.exports = {
+  experimental: {
+    amp: {
+      validator: require.resolve('./amp-validator-wasm.js'),
+    },
+  },
+}
