+++ test/e2e/async-modules/pages/_app.jsx
@@ -1,0 +1,5 @@
+const appValue = await Promise.resolve('hello')
+
+export default function MyApp({ Component, pageProps }) {
+  return <Component {...pageProps} appValue={appValue} />
+}
