+++ test/e2e/async-modules/pages/_error.jsx
@@ -1,0 +1,7 @@
+const errorContent = await Promise.resolve('hello error')
+
+function Error({ statusCode }) {
+  return <p id="content-error">{errorContent}</p>
+}
+
+export default Error
