+++ test/e2e/async-modules/pages/api/hello.js
@@ -1,0 +1,5 @@
+const value = await Promise.resolve(42)
+
+export default function (req, res) {
+  res.json({ value })
+}
