+++ test/e2e/async-modules/pages/gssp.jsx
@@ -1,0 +1,15 @@
+const gsspValue = await Promise.resolve(42)
+
+export async function getServerSideProps() {
+  return {
+    props: { gsspValue },
+  }
+}
+
+export default function Index({ gsspValue }) {
+  return (
+    <main>
+      <div id="gssp-value">{gsspValue}</div>
+    </main>
+  )
+}
