+++ test/e2e/async-modules/pages/make-error.jsx
@@ -1,0 +1,7 @@
+export async function getServerSideProps() {
+  throw new Error('BOOM')
+}
+
+export default function Page() {
+  return <div>hello</div>
+}
