--- test/integration/async-modules/pages/404.jsx
@@ -1,5 +1,0 @@
-const content = await Promise.resolve("hi y'all")
-
-export default function Custom404() {
-  return <h1 id="content-404">{content}</h1>
-}
