--- test/integration/async-modules/pages/_app.jsx
@@ -1,5 +1,0 @@
-const appValue = await Promise.resolve('hello')
-
-export default function MyApp({ Component, pageProps }) {
-  return <Component {...pageProps} appValue={appValue} />
-}
