--- test/integration/async-modules/pages/_error.jsx
@@ -1,7 +1,0 @@
-const errorContent = await Promise.resolve('hello error')
-
-function Error({ statusCode }) {
-  return <p id="content-error">{errorContent}</p>
-}
-
-export default Error
