--- test/integration/async-modules/pages/api/hello.js
@@ -1,5 +1,0 @@
-const value = await Promise.resolve(42)
-
-export default function (req, res) {
-  res.json({ value })
-}
