--- test/integration/async-modules/pages/make-error.jsx
@@ -1,7 +1,0 @@
-export async function getServerSideProps() {
-  throw new Error('BOOM')
-}
-
-export default function Page() {
-  return <div>hello</div>
-}
